Bug 951950 Fix tests that don't run on Fennec, fix event/core iterating over invalid properties (localStorage) for Fennec's tab which accepts a window
