+++ lib/sdk/event/core.js
@@ -161,9 +161,10 @@ function setListeners(target, listeners) {
   Object.keys(listeners || {}).forEach(key => {
     let match = EVENT_TYPE_PATTERN.exec(key);
     let type = match && match[1].toLowerCase();
-    let listener = listeners[key];
+    if (!type) return;
 
-    if (type && typeof(listener) === 'function')
+    let listener = listeners[key];
+    if (typeof(listener) === 'function')
       on(target, type, listener);
   });
 }
